Move test to persistence test file
